+++ ts-sdk/core/tests/types.test.ts
@@ -10,8 +10,13 @@
 
 import { describe, it } from "vitest";
 
+<<<<<<< HEAD
 import type { FusionPool, Position, TickArray } from "../../client/src";
 import type { FusionPoolFacade, PositionFacade, TickArrayFacade } from "../dist/nodejs/fusionamm_core_js_bindings";
+=======
+import type { FusionPool, Position } from "../../client/src";
+import type { FusionPoolFacade, PositionFacade } from "../dist/nodejs/fusionamm_core_js_bindings";
+>>>>>>> upstream/main
 
 // Since these tests are only for type checking, nothing actually happens at runtime.
 
@@ -26,8 +31,8 @@ describe("WASM exported types match Codama types", () => {
     fauxPosition satisfies PositionFacade;
   });
 
-  it("TickArray", () => {
-    const fauxTickArray = {} as TickArray;
-    fauxTickArray satisfies TickArrayFacade;
-  });
+  //  it("TickArray", () => {
+  //    const fauxTickArray = {} as TickArray;
+  //    fauxTickArray satisfies TickArrayFacade;
+  //  });
 });