+++ ts-sdk/fusionamm-sdk/src/index.ts
@@ -21,3 +21,7 @@ export * from "./swap";
 export * from "./collectProtocolFees";
 export * from "./resetPoolPrice";
 export * from "./lockPosition";
+<<<<<<< HEAD
+=======
+export * from "./sysvar";
+>>>>>>> upstream/main