+++ rust-sdk/client/CHANGELOG.md
@@ -1,10 +1,26 @@
 # @crypticdot/fusionamm-rust-client
 
+## 1.1.5
+
+### Patch Changes
+
+- 89cb3b5: Fixed typo in fetch_all_tick_array_with_filter
+
+## 1.1.4
+
+### Patch Changes
+
+- 4423944: Removed obsolete tick array account
+
+## 1.1.3
+
+## 1.1.2
+
 ## 1.1.1
 
 ### Patch Changes
 
-- 60b8753: Made tick array structures publically visible
+- 60b8753: Made tick array structures publicly visible
 
 ## 1.1.0