+++ rust-sdk/fusionamm-sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crypticdot/fusionamm-rust
 
+## 1.1.5
+
+## 1.1.4
+
+### Patch Changes
+
+- 4423944: Removed obsolete tick array account
+
+## 1.1.3
+
+## 1.1.2
+
 ## 1.1.1
 
 ## 1.1.0