+++ ts-sdk/core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @crypticdot/fusionamm-core
 
+<<<<<<< HEAD
+=======
+## 1.1.5
+
+## 1.1.4
+
+### Patch Changes
+
+- 4423944: Removed obsolete tick array account
+
+## 1.1.3
+
+## 1.1.2
+
+>>>>>>> upstream/main
 ## 1.1.1
 
 ## 1.1.0